End Points Peliculas
Se crearon los end points para Pelicula

+++ src/main/java/com/example/cinemagicgto/cinemarestful/repository/PeliculaRepository.java
@@ -1,9 +1,11 @@
 package com.example.cinemagicgto.cinemarestful.repository;
 
-import com.example.cinemagicgto.cinemarestful.data.Pelicula;
+import com.example.cinemagicgto.cinemarestful.entity.Pelicula;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PeliculaRepository  extends JpaRepository<Pelicula, Long> {
+public interface PeliculaRepository  extends JpaRepository<Pelicula, Integer> {
+
+
 
 
 }
